fix(experimental): scale grid index dispatches
